picogame: one file per type, reuse the core error messages
Review feedback. The types that were consolidated in __init__.c (Bitmap,
Sprite, StripDraw, Triangles, Framebuffer) each get their own shared-bindings
file, matching Canvas/Scene/Tilemap/Particles/Display, so __init__.c is the
module level only: its docstring, the module functions and the globals table.
The docstring now also says how picogame relates to displayio - same display
object, different way to drive it, no retained pixels - which is the first
thing a reader of that file should learn.

The eight error messages the module added are replaced with ones CircuitPython
already ships (mp_arg_validate_type / mp_arg_error_invalid /
mp_arg_validate_length_min, and m_malloc_fail for the unreachable scene-cap
guard): the module now contributes no new strings to the translations.
