+++ locale/circuitpython.pot
@@ -1571,14 +1571,6 @@ msgstr ""
 msgid "Number of data_pins must be %d or %d, not %d"
 msgstr ""
 
-#: ports/espressif/common-hal/picogame/Display.c
-msgid "rgb444 fast Display not supported on this port yet"
-msgstr ""
-
-#: ports/espressif/common-hal/picogame/Display.c
-msgid "fast Display needs a FourWire SPI display"
-msgstr ""
-
 #: ports/espressif/common-hal/pulseio/PulseIn.c
 msgid "pop from an empty PulseIn"
 msgstr ""
@@ -1823,7 +1815,7 @@ msgstr ""
 
 #: ports/raspberrypi/bindings/cyw43/__init__.c py/argcheck.c py/objexcept.c
 #: shared-bindings/bitmapfilter/__init__.c shared-bindings/canio/CAN.c
-#: shared-bindings/digitalio/Pull.c shared-bindings/picogame/__init__.c
+#: shared-bindings/digitalio/Pull.c shared-bindings/picogame/Sprite.c
 #: shared-bindings/supervisor/__init__.c shared-module/audiofilters/Filter.c
 #: shared-module/displayio/__init__.c shared-module/synthio/Synthesizer.c
 msgid "%q must be of type %q or %q, not %q"
@@ -2890,9 +2882,9 @@ msgstr ""
 msgid "schedule queue full"
 msgstr ""
 
-#: py/modstruct.c shared-bindings/picogame/Canvas.c
-#: shared-bindings/picogame/__init__.c shared-module/picogame/__init__.c
-#: shared-module/struct/__init__.c
+#: py/modstruct.c shared-bindings/picogame/Bitmap.c
+#: shared-bindings/picogame/Canvas.c shared-bindings/picogame/Framebuffer.c
+#: shared-module/picogame/__init__.c shared-module/struct/__init__.c
 msgid "buffer too small"
 msgstr ""
 
@@ -4100,34 +4092,9 @@ msgstr ""
 msgid "Specify exactly one of data0 or data_pins"
 msgstr ""
 
-#: shared-bindings/picogame/Display.c shared-bindings/picogame/Scene.c
-#: shared-bindings/picogame/__init__.c
-msgid "expected a BusDisplay"
-msgstr ""
-
-#: shared-bindings/picogame/Scene.c
-msgid "scene full"
-msgstr ""
-
-#: shared-bindings/picogame/Scene.c
-msgid "item not in scene"
-msgstr ""
-
-#: shared-bindings/picogame/__init__.c shared-module/msgpack/__init__.c
+#: shared-bindings/picogame/Bitmap.c shared-module/msgpack/__init__.c
 #: supervisor/shared/settings.c
 msgid "Invalid format"
-msgstr ""
-
-#: shared-bindings/picogame/__init__.c
-msgid "PAL8 needs a palette"
-msgstr ""
-
-#: shared-bindings/picogame/__init__.c
-msgid "palette is empty"
-msgstr ""
-
-#: shared-bindings/picogame/__init__.c
-msgid "expected a Sprite, Tilemap, Particles, Canvas, StripDraw or Triangles"
 msgstr ""
 
 #: shared-bindings/ps2io/Ps2.c
