+++ shared-bindings/picogame/__init__.c
@@ -23,290 +23,15 @@
 #include "shared-bindings/picogame/Particles.h"
 #include "shared-bindings/picogame/Canvas.h"
 #include "shared-bindings/picogame/Framebuffer.h"
+#include "shared-bindings/picogame/StripDraw.h"
+#include "shared-bindings/picogame/Triangles.h"
 #include "shared-module/picogame/__init__.h"
 #include "shared-module/picogame/Bitmap.h"
 #include "shared-module/picogame/Sprite.h"
 
 // ---------------------------------------------------------------------------
-// Bitmap
+// Module-level functions
 // ---------------------------------------------------------------------------
-
-//| class Bitmap:
-//|     """An image atlas of one or more equal-size frames, of arbitrary size.
-//|
-//|     Unlike ``_stage`` (fixed 16x16 tiles), frames may be any width/height.
-//|     Pixel data and palette entries must be in the display's wire byte order
-//|     (use :py:func:`picogame.rgb565` to build colors)."""
-//|
-//|     def __init__(
-//|         self,
-//|         data: ReadableBuffer,
-//|         width: int,
-//|         height: int,
-//|         *,
-//|         format: int = RGB565,
-//|         palette: Optional[ReadableBuffer] = None,
-//|         frames: int = 1,
-//|         stride: int = 0,
-//|         transparent: Optional[int] = None,
-//|     ) -> None: ...
-//|
-// Int pixel/scale -> 24.8 fixed-point. Shift through unsigned so a wild coordinate (e.g. 100_000_000,
-// in 32-bit mp_int range but not after <<8) wraps modularly instead of hitting signed-overflow UB.
-// Costs nothing over a plain shift; for any real on-screen coordinate the result is identical.
-static int32_t pg_int_to_fp8(mp_int_t v) {
-    return (int32_t)((uint32_t)v << 8);
-}
-
-static mp_obj_t picogame_bitmap_make_new(const mp_obj_type_t *type, size_t n_args,
-    size_t n_kw, const mp_obj_t *all_args) {
-    enum { ARG_data, ARG_width, ARG_height, ARG_format, ARG_palette, ARG_frames, ARG_stride, ARG_transparent };
-    static const mp_arg_t allowed_args[] = {
-        { MP_QSTR_data, MP_ARG_REQUIRED | MP_ARG_OBJ },
-        { MP_QSTR_width, MP_ARG_REQUIRED | MP_ARG_INT },
-        { MP_QSTR_height, MP_ARG_REQUIRED | MP_ARG_INT },
-        { MP_QSTR_format, MP_ARG_KW_ONLY | MP_ARG_INT, {.u_int = PICOGAME_FMT_RGB565} },
-        { MP_QSTR_palette, MP_ARG_KW_ONLY | MP_ARG_OBJ, {.u_obj = mp_const_none} },
-        { MP_QSTR_frames, MP_ARG_KW_ONLY | MP_ARG_INT, {.u_int = 1} },
-        { MP_QSTR_stride, MP_ARG_KW_ONLY | MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_transparent, MP_ARG_KW_ONLY | MP_ARG_OBJ, {.u_obj = mp_const_none} },
-    };
-    mp_arg_val_t args[MP_ARRAY_SIZE(allowed_args)];
-    mp_arg_parse_all_kw_array(n_args, n_kw, all_args, MP_ARRAY_SIZE(allowed_args), allowed_args, args);
-
-    // bound BEFORE any arithmetic: an unbounded width could overflow width*frames (int32) and wrap
-    // small, slipping past the size guards below into an undersized buffer (OOB read in the blitter).
-    mp_int_t width = mp_arg_validate_int_range(args[ARG_width].u_int, 1, 65535, MP_QSTR_width);
-    mp_int_t height = mp_arg_validate_int_range(args[ARG_height].u_int, 1, 65535, MP_QSTR_height);
-    mp_int_t frames = mp_arg_validate_int_range(args[ARG_frames].u_int, 1, 255, MP_QSTR_frames);
-    mp_int_t format = args[ARG_format].u_int;
-    if (format != PICOGAME_FMT_RGB565 && format != PICOGAME_FMT_PAL8) {
-        mp_raise_ValueError(MP_ERROR_TEXT("Invalid format"));
-    }
-    mp_int_t stride = args[ARG_stride].u_int;
-    if (stride <= 0) {
-        stride = width * frames;
-    }
-    // stride must hold the whole horizontal atlas, and the dims must fit the uint16_t fields,
-    // or blits index past a row / the stored stride truncates.
-    mp_arg_validate_int_max(width * frames, 65535, MP_QSTR_width);
-    mp_arg_validate_int_max(stride, 65535, MP_QSTR_stride);
-    mp_arg_validate_int_min(stride, width * frames, MP_QSTR_stride);
-
-    mp_buffer_info_t data_info;
-    mp_get_buffer_raise(args[ARG_data].u_obj, &data_info, MP_BUFFER_READ);
-
-    const uint16_t *palette = NULL;
-    mp_obj_t palette_obj = MP_OBJ_NULL;
-    size_t pal_len = 0;
-    if (format == PICOGAME_FMT_PAL8) {
-        if (args[ARG_palette].u_obj == mp_const_none) {
-            mp_raise_ValueError(MP_ERROR_TEXT("PAL8 needs a palette"));
-        }
-        mp_buffer_info_t pal_info;
-        mp_get_buffer_raise(args[ARG_palette].u_obj, &pal_info, MP_BUFFER_READ);
-        palette = pal_info.buf;
-        palette_obj = args[ARG_palette].u_obj;
-        pal_len = pal_info.len;
-    }
-
-    size_t bpp = (format == PICOGAME_FMT_PAL8) ? 1 : 2;
-    // 64-bit: stride*height*bpp can exceed 32 bits (stride,height <= 65535) and wrap small in a 32-bit
-    // size_t, letting a tiny buffer pass this check -> OOB read in the blitter. Compute + compare wide.
-    uint64_t need = (uint64_t)stride * (uint64_t)height * (uint64_t)bpp;
-    if ((uint64_t)data_info.len < need) {
-        mp_raise_ValueError(MP_ERROR_TEXT("buffer too small"));
-    }
-    if (format == PICOGAME_FMT_PAL8 && pal_len < 2) {
-        // Need >=1 entry so pal[0] is valid. Contract (see blitter): PAL8 indices MUST be < palette
-        // length. The caller may write indices into the (mutable) `data` buffer directly, so we cannot
-        // validate once here. An out-of-range index is undefined behaviour (reads past the palette:
-        // usually a garbage colour, but it may fault on some platforms). Per-pixel clamping was dropped
-        // for speed - see the "restore full bounds-safety" note in the PAL8 blit loop if it's needed.
-        mp_raise_ValueError(MP_ERROR_TEXT("palette is empty"));
-    }
-
-    picogame_bitmap_obj_t *self = mp_obj_malloc(picogame_bitmap_obj_t, type);
-    self->data_obj = args[ARG_data].u_obj;
-    self->palette_obj = palette_obj;
-    self->data = data_info.buf;
-    self->palette = palette;
-    self->width = width;
-    self->height = height;
-    self->stride = stride;
-    self->frames = frames;
-    self->format = format;
-    // palette length in entries (informational; blitter assumes indices < this - see blit contract).
-    self->pal_entries = (uint16_t)((pal_len / 2) > 65535 ? 65535 : (pal_len / 2));
-    if (args[ARG_transparent].u_obj != mp_const_none) {
-        self->transparent = mp_obj_get_int(args[ARG_transparent].u_obj);
-        self->has_transparent = true;
-    } else {
-        self->transparent = 0;
-        self->has_transparent = false;
-    }
-    return MP_OBJ_FROM_PTR(self);
-}
-
-//|
-//|     width: int
-//|     height: int
-//|     frames: int
-//|     """Frame dimensions and frame count (read-only)."""
-static mp_obj_t bitmap_get_width(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_bitmap_obj_t *)MP_OBJ_TO_PTR(self_in))->width);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_width_obj, bitmap_get_width);
-MP_PROPERTY_GETTER(bitmap_width_obj, (mp_obj_t)&bitmap_get_width_obj);
-
-static mp_obj_t bitmap_get_height(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_bitmap_obj_t *)MP_OBJ_TO_PTR(self_in))->height);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_height_obj, bitmap_get_height);
-MP_PROPERTY_GETTER(bitmap_height_obj, (mp_obj_t)&bitmap_get_height_obj);
-
-static mp_obj_t bitmap_get_frames(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_bitmap_obj_t *)MP_OBJ_TO_PTR(self_in))->frames);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_frames_obj, bitmap_get_frames);
-MP_PROPERTY_GETTER(bitmap_frames_obj, (mp_obj_t)&bitmap_get_frames_obj);
-
-//|     format: int
-//|     """RGB565 or PAL8 (read-only)."""
-static mp_obj_t bitmap_get_format(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_bitmap_obj_t *)MP_OBJ_TO_PTR(self_in))->format);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_format_obj, bitmap_get_format);
-MP_PROPERTY_GETTER(bitmap_format_obj, (mp_obj_t)&bitmap_get_format_obj);
-
-//|     stride: int
-//|     """Row stride in pixels (read-only)."""
-static mp_obj_t bitmap_get_stride(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_bitmap_obj_t *)MP_OBJ_TO_PTR(self_in))->stride);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_stride_obj, bitmap_get_stride);
-MP_PROPERTY_GETTER(bitmap_stride_obj, (mp_obj_t)&bitmap_get_stride_obj);
-
-//|     palette: Optional[ReadableBuffer]
-//|     """The PAL8 palette buffer this Bitmap was built with, or None for RGB565
-//|     (read-only). Lets palette helpers read it back instead of holding a sidecar ref."""
-static mp_obj_t bitmap_get_palette(mp_obj_t self_in) {
-    picogame_bitmap_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    return (self->palette_obj == MP_OBJ_NULL) ? mp_const_none : self->palette_obj;
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_palette_obj, bitmap_get_palette);
-MP_PROPERTY_GETTER(bitmap_palette_obj, (mp_obj_t)&bitmap_get_palette_obj);
-
-//|     transparent: Optional[int]
-//|     """The transparent color/index, or None if the Bitmap is fully opaque (read-only)."""
-//|
-//|
-static mp_obj_t bitmap_get_transparent(mp_obj_t self_in) {
-    picogame_bitmap_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    return self->has_transparent ? MP_OBJ_NEW_SMALL_INT(self->transparent) : mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(bitmap_get_transparent_obj, bitmap_get_transparent);
-MP_PROPERTY_GETTER(bitmap_transparent_obj, (mp_obj_t)&bitmap_get_transparent_obj);
-
-static const mp_rom_map_elem_t picogame_bitmap_locals_dict_table[] = {
-    { MP_ROM_QSTR(MP_QSTR_width), MP_ROM_PTR(&bitmap_width_obj) },
-    { MP_ROM_QSTR(MP_QSTR_height), MP_ROM_PTR(&bitmap_height_obj) },
-    { MP_ROM_QSTR(MP_QSTR_frames), MP_ROM_PTR(&bitmap_frames_obj) },
-    { MP_ROM_QSTR(MP_QSTR_format), MP_ROM_PTR(&bitmap_format_obj) },
-    { MP_ROM_QSTR(MP_QSTR_stride), MP_ROM_PTR(&bitmap_stride_obj) },
-    { MP_ROM_QSTR(MP_QSTR_palette), MP_ROM_PTR(&bitmap_palette_obj) },
-    { MP_ROM_QSTR(MP_QSTR_transparent), MP_ROM_PTR(&bitmap_transparent_obj) },
-};
-static MP_DEFINE_CONST_DICT(picogame_bitmap_locals_dict, picogame_bitmap_locals_dict_table);
-
-MP_DEFINE_CONST_OBJ_TYPE(
-    picogame_bitmap_type,
-    MP_QSTR_Bitmap,
-    MP_TYPE_FLAG_HAS_SPECIAL_ACCESSORS,
-    make_new, picogame_bitmap_make_new,
-    locals_dict, &picogame_bitmap_locals_dict
-    );
-
-// ---------------------------------------------------------------------------
-// Sprite
-// ---------------------------------------------------------------------------
-
-//| class Sprite:
-//|     """A positioned, animatable instance of a :py:class:`Bitmap`."""
-//|
-//|     def __init__(
-//|         self,
-//|         bitmap: Bitmap,
-//|         x: int = 0,
-//|         y: int = 0,
-//|         *,
-//|         frame: int = 0,
-//|         visible: bool = True,
-//|         flip_x: bool = False,
-//|         flip_y: bool = False,
-//|     ) -> None: ...
-//|
-static mp_obj_t picogame_sprite_make_new(const mp_obj_type_t *type, size_t n_args,
-    size_t n_kw, const mp_obj_t *all_args) {
-    enum { ARG_bitmap, ARG_x, ARG_y, ARG_frame, ARG_visible, ARG_flip_x, ARG_flip_y };
-    static const mp_arg_t allowed_args[] = {
-        { MP_QSTR_bitmap, MP_ARG_REQUIRED | MP_ARG_OBJ },
-        { MP_QSTR_x, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_y, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_frame, MP_ARG_KW_ONLY | MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_visible, MP_ARG_KW_ONLY | MP_ARG_BOOL, {.u_bool = true} },
-        { MP_QSTR_flip_x, MP_ARG_KW_ONLY | MP_ARG_BOOL, {.u_bool = false} },
-        { MP_QSTR_flip_y, MP_ARG_KW_ONLY | MP_ARG_BOOL, {.u_bool = false} },
-    };
-    mp_arg_val_t args[MP_ARRAY_SIZE(allowed_args)];
-    mp_arg_parse_all_kw_array(n_args, n_kw, all_args, MP_ARRAY_SIZE(allowed_args), allowed_args, args);
-
-    mp_obj_t bitmap_obj = mp_arg_validate_type(args[ARG_bitmap].u_obj, &picogame_bitmap_type, MP_QSTR_bitmap);
-
-    picogame_sprite_obj_t *self = mp_obj_malloc(picogame_sprite_obj_t, type);
-    self->bitmap = MP_OBJ_TO_PTR(bitmap_obj);
-    self->x = pg_int_to_fp8(args[ARG_x].u_int);   // pixel -> 24.8 fixed-point (overflow-clamped)
-    self->y = pg_int_to_fp8(args[ARG_y].u_int);
-    self->frame = args[ARG_frame].u_int;
-    self->flags = (args[ARG_visible].u_bool ? PICOGAME_SPR_VISIBLE : 0)
-        | (args[ARG_flip_x].u_bool ? PICOGAME_SPR_FLIP_X : 0)
-        | (args[ARG_flip_y].u_bool ? PICOGAME_SPR_FLIP_Y : 0);
-    self->anchor_x = 0;   // default pivot = top-left (0, 0)
-    self->anchor_y = 0;
-    self->scale = 256;    // 8.8 fixed-point: 256 = 1.0x (no scaling)
-    self->angle = 0;      // no rotation -> axis-aligned fast path
-    self->flash_color = 0;
-    self->dither = 0;
-    self->data = mp_const_none;
-    self->seq = 0;            // dirty-rect change counter (mp_obj_malloc zeroes anyway; explicit)
-    return MP_OBJ_FROM_PTR(self);
-}
-
-static void set_flag(picogame_sprite_obj_t *self, uint8_t flag, bool on) {
-    if (on) {
-        self->flags |= flag;
-    } else {
-        self->flags &= ~flag;
-    }
-}
-
-// shadow / flash / dither are mutually exclusive (one blit effect at a time); setting one ON
-// clears the others, so "the last effect you set wins". Turning one OFF clears ONLY its own flag,
-// so clearing an effect you never enabled (e.g. spr.flash = 0) can't wipe a different active one.
-#define PICOGAME_SPR_FX_MASK (PICOGAME_SPR_SHADOW | PICOGAME_SPR_FLASH | PICOGAME_SPR_DITHER | PICOGAME_SPR_TINT)
-static void set_effect(picogame_sprite_obj_t *self, uint8_t flag, bool on) {
-    if (on) {
-        self->flags = (uint8_t)((self->flags & ~PICOGAME_SPR_FX_MASK) | flag);
-    } else {
-        self->flags &= (uint8_t) ~flag;
-    }
-}
-
-// Round a float to 24.8 fixed-point (shared by the position/scale/anchor setters so the soft-float
-// round sequence is emitted once, not per call site).
-static int32_t pg_round_fp8(mp_float_t f) {
-    return (int32_t)(f * 256 + (f >= 0 ? (mp_float_t)0.5 : (mp_float_t)-0.5));
-}
 
 // Cast a (possibly subclassed) BusDisplay arg to its native object, raising if it isn't one.
 // Also accepts the pg.Display fast-DMA wrapper (unwrapped to its underlying busdisplay - the
@@ -322,758 +47,35 @@ static busdisplay_busdisplay_obj_t *pg_get_display(mp_obj_t obj) {
     #endif
     mp_obj_t native = mp_obj_cast_to_native_base(obj, &busdisplay_busdisplay_type);
     if (!mp_obj_is_type(native, &busdisplay_busdisplay_type)) {
-        mp_raise_TypeError(MP_ERROR_TEXT("expected a BusDisplay"));
+        mp_arg_validate_type(native, &busdisplay_busdisplay_type, MP_QSTR_display);
     }
     return MP_OBJ_TO_PTR(native);
 }
-
-// Accept an int or float and store as 24.8 fixed-point (rounded). Integer fast path avoids
-// software float on RP2040 - game code sets x/y (via move/setters) every frame, usually with ints.
-static int32_t obj_to_fp(mp_obj_t o) {
-    if (mp_obj_is_int(o)) {
-        return pg_int_to_fp8(mp_obj_get_int(o));
-    }
-    mp_float_t f = mp_obj_get_float(o);
-    return pg_round_fp8(f);
-}
-
-//|
-//|     x: int
-//|     y: int
-//|     """Integer pixel position (scene coords). Setting accepts a float for
-//|     sub-pixel placement; reading returns the floored pixel."""
-//|     fx: float
-//|     fy: float
-//|     """Sub-pixel position (use for smooth physics: e.g. ``sprite.fx += 2.4``)."""
-// FLASH NOTE (property objects vs. a single `attr` handler): Sprite (and the other property-dense
-// types) expose each attribute as its own getter/setter + MP_PROPERTY object below - the standard
-// CircuitPython shared-bindings idiom (~423 such uses across CP). Collapsing these onto ONE `attr`
-// load/store function per type (a qstr switch, like py/objcomplex.c's complex_attr for .real/.imag)
-// would save ~3-4 KB of flash across the whole binding layer (Sprite alone ~1.5-2.5 KB). It is a
-// valid MicroPython mechanism and would NOT break the .pyi stubs (those are generated from the //|
-// comments, not the C property objects). We deliberately KEEP property objects: they are the
-// idiomatic, most readable shared-bindings form and keep every type's definition uniform. Revisit
-// only if flash becomes critical (then convert Sprite first - the densest cluster - not every type).
-static mp_obj_t sprite_get_x(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->x >> 8);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_x_obj, sprite_get_x);
-static mp_obj_t sprite_set_x(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->x = obj_to_fp(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_x_obj, sprite_set_x);
-MP_PROPERTY_GETSET(sprite_x_obj, (mp_obj_t)&sprite_get_x_obj, (mp_obj_t)&sprite_set_x_obj);
-
-static mp_obj_t sprite_get_y(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->y >> 8);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_y_obj, sprite_get_y);
-static mp_obj_t sprite_set_y(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->y = obj_to_fp(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_y_obj, sprite_set_y);
-MP_PROPERTY_GETSET(sprite_y_obj, (mp_obj_t)&sprite_get_y_obj, (mp_obj_t)&sprite_set_y_obj);
-
-static mp_obj_t sprite_get_fx(mp_obj_t self_in) {
-    return mp_obj_new_float(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->x * (mp_float_t)(1.0 / 256.0));
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_fx_obj, sprite_get_fx);
-// set_fx is byte-identical to set_x (both store obj_to_fp(v) into ->x) -> reuse set_x's fun obj.
-MP_PROPERTY_GETSET(sprite_fx_obj, (mp_obj_t)&sprite_get_fx_obj, (mp_obj_t)&sprite_set_x_obj);
-
-static mp_obj_t sprite_get_fy(mp_obj_t self_in) {
-    return mp_obj_new_float(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->y * (mp_float_t)(1.0 / 256.0));
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_fy_obj, sprite_get_fy);
-// set_fy is byte-identical to set_y -> reuse set_y's fun obj.
-MP_PROPERTY_GETSET(sprite_fy_obj, (mp_obj_t)&sprite_get_fy_obj, (mp_obj_t)&sprite_set_y_obj);
-
-static mp_obj_t sprite_get_frame(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->frame);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_frame_obj, sprite_get_frame);
-static mp_obj_t sprite_set_frame(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->frame = mp_obj_get_int(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_frame_obj, sprite_set_frame);
-MP_PROPERTY_GETSET(sprite_frame_obj, (mp_obj_t)&sprite_get_frame_obj, (mp_obj_t)&sprite_set_frame_obj);
-
-static mp_obj_t sprite_get_visible(mp_obj_t self_in) {
-    return mp_obj_new_bool((((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->flags & PICOGAME_SPR_VISIBLE) != 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_visible_obj, sprite_get_visible);
-static mp_obj_t sprite_set_visible(mp_obj_t self_in, mp_obj_t v) {
-    set_flag(MP_OBJ_TO_PTR(self_in), PICOGAME_SPR_VISIBLE, mp_obj_is_true(v));
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_visible_obj, sprite_set_visible);
-MP_PROPERTY_GETSET(sprite_visible_obj, (mp_obj_t)&sprite_get_visible_obj, (mp_obj_t)&sprite_set_visible_obj);
-
-static mp_obj_t sprite_get_flip_x(mp_obj_t self_in) {
-    return mp_obj_new_bool((((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->flags & PICOGAME_SPR_FLIP_X) != 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_flip_x_obj, sprite_get_flip_x);
-static mp_obj_t sprite_set_flip_x(mp_obj_t self_in, mp_obj_t v) {
-    set_flag(MP_OBJ_TO_PTR(self_in), PICOGAME_SPR_FLIP_X, mp_obj_is_true(v));
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_flip_x_obj, sprite_set_flip_x);
-MP_PROPERTY_GETSET(sprite_flip_x_obj, (mp_obj_t)&sprite_get_flip_x_obj, (mp_obj_t)&sprite_set_flip_x_obj);
-
-static mp_obj_t sprite_get_flip_y(mp_obj_t self_in) {
-    return mp_obj_new_bool((((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->flags & PICOGAME_SPR_FLIP_Y) != 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_flip_y_obj, sprite_get_flip_y);
-static mp_obj_t sprite_set_flip_y(mp_obj_t self_in, mp_obj_t v) {
-    set_flag(MP_OBJ_TO_PTR(self_in), PICOGAME_SPR_FLIP_Y, mp_obj_is_true(v));
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_flip_y_obj, sprite_set_flip_y);
-MP_PROPERTY_GETSET(sprite_flip_y_obj, (mp_obj_t)&sprite_get_flip_y_obj, (mp_obj_t)&sprite_set_flip_y_obj);
-
-//|     scale: float
-//|     """Uniform draw scale (nearest-neighbour). 1.0 = native (fast path); 2.0 = double
-//|     size, fractional values are allowed (e.g. a powerup grow tween). Anchor stays put."""
-static mp_obj_t sprite_get_scale(mp_obj_t self_in) {
-    return mp_obj_new_float(((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->scale * (mp_float_t)(1.0 / 256.0));
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_scale_obj, sprite_get_scale);
-static mp_obj_t sprite_set_scale(mp_obj_t self_in, mp_obj_t v) {
-    int q;
-    if (mp_obj_is_int(v)) {                          // int fast path (no software float on RP2040)
-        q = pg_int_to_fp8(mp_obj_get_int(v));        // overflow-clamped <<8
-    } else {
-        mp_float_t f = mp_obj_get_float(v);
-        q = pg_round_fp8(f);
-    }
-    if (q < 1) {
-        q = 1;
-    }
-    if (q > 65535) {
-        q = 65535;
-    }
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    self->scale = (uint16_t)q;
-    self->xf_valid = 0;                              // affine cache depends on scale
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_scale_obj, sprite_set_scale);
-MP_PROPERTY_GETSET(sprite_scale_obj, (mp_obj_t)&sprite_get_scale_obj, (mp_obj_t)&sprite_set_scale_obj);
-
-//|     angle: float
-//|     """Rotation in degrees about the anchor (0 = none, the fast path). Nearest-neighbour,
-//|     so integer scales stay crisp; rotation shimmers slightly (pixel-art trade-off)."""
-static mp_obj_t sprite_get_angle(mp_obj_t self_in) {
-    return mp_obj_new_float((mp_float_t)((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->angle);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_angle_obj, sprite_get_angle);
-static mp_obj_t sprite_set_angle(mp_obj_t self_in, mp_obj_t v) {
-    int a;
-    if (mp_obj_is_int(v)) {                          // int fast path (no software float on RP2040)
-        a = mp_obj_get_int(v);
-    } else {
-        mp_float_t f = mp_obj_get_float(v);
-        a = (int)(f + (f >= 0 ? (mp_float_t)0.5 : (mp_float_t)-0.5));
-    }
-    a %= 360;
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    self->angle = (int16_t)a;
-    self->xf_valid = 0;                              // affine cache depends on angle
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_angle_obj, sprite_set_angle);
-MP_PROPERTY_GETSET(sprite_angle_obj, (mp_obj_t)&sprite_get_angle_obj, (mp_obj_t)&sprite_set_angle_obj);
-
-//|     shadow: bool
-//|     """Draw opaque pixels as a darkened destination instead of colour - a drop-shadow
-//|     silhouette (offset copy below the sprite) or a dim overlay (a solid sprite scaled
-//|     over a dialog/pause area)."""
-static mp_obj_t sprite_get_shadow(mp_obj_t self_in) {
-    return mp_obj_new_bool((((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->flags & PICOGAME_SPR_SHADOW) != 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_shadow_obj, sprite_get_shadow);
-static mp_obj_t sprite_set_shadow(mp_obj_t self_in, mp_obj_t v) {
-    set_effect(MP_OBJ_TO_PTR(self_in), PICOGAME_SPR_SHADOW, mp_obj_is_true(v));
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_shadow_obj, sprite_set_shadow);
-MP_PROPERTY_GETSET(sprite_shadow_obj, (mp_obj_t)&sprite_get_shadow_obj, (mp_obj_t)&sprite_set_shadow_obj);
-
-//|     flash: int
-//|     """Draw opaque pixels as a solid colour (a wire-order RGB565 int from rgb565) instead
-//|     of their own colour - a hit-flash or tint. Set to a colour to enable, 0/False to turn
-//|     off. Pulse it for 1-3 frames on impact. Mutually exclusive with shadow/dither."""
-static mp_obj_t sprite_get_flash(mp_obj_t self_in) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    return MP_OBJ_NEW_SMALL_INT((s->flags & PICOGAME_SPR_FLASH) ? s->flash_color : 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_flash_obj, sprite_get_flash);
-static mp_obj_t sprite_set_flash(mp_obj_t self_in, mp_obj_t v) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    // Falsy (None / False / 0) turns flash OFF. A non-zero colour enables it. (You can't flash
-    // pure black - use a near-black colour if you ever need that; off is the common case.)
-    if (!mp_obj_is_true(v)) {
-        set_effect(s, PICOGAME_SPR_FLASH, false);
-    } else {
-        s->flash_color = (uint16_t)mp_obj_get_int(v);
-        set_effect(s, PICOGAME_SPR_FLASH, true);
-    }
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_flash_obj, sprite_set_flash);
-MP_PROPERTY_GETSET(sprite_flash_obj, (mp_obj_t)&sprite_get_flash_obj, (mp_obj_t)&sprite_set_flash_obj);
-
-//|     dither: int
-//|     """Fake transparency via an ordered (Bayer) dither, no alpha blending: 0 = opaque
-//|     (off), 8 = ~50% see-through, 16 = invisible. A classic 1-bit look - for ghosts,
-//|     fading/spawning enemies, fog, force fields. Mutually exclusive with shadow/flash."""
-static mp_obj_t sprite_get_dither(mp_obj_t self_in) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    return MP_OBJ_NEW_SMALL_INT((s->flags & PICOGAME_SPR_DITHER) ? s->dither : 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_dither_obj, sprite_get_dither);
-static mp_obj_t sprite_set_dither(mp_obj_t self_in, mp_obj_t v) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    int lv = mp_obj_get_int(v);
-    if (lv < 0) {
-        lv = 0;
-    }
-    if (lv > 16) {
-        lv = 16;
-    }
-    s->dither = (uint8_t)lv;
-    set_effect(s, PICOGAME_SPR_DITHER, lv > 0);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_dither_obj, sprite_set_dither);
-MP_PROPERTY_GETSET(sprite_dither_obj, (mp_obj_t)&sprite_get_dither_obj, (mp_obj_t)&sprite_set_dither_obj);
-
-//|     tint: int
-//|     """Multiply opaque pixels by a colour (wire-order RGB565 from rgb565), keeping the
-//|     sprite's shading - coloured lighting, a red damage flush, a blue freeze, a power-up
-//|     glow. Unlike ``flash`` (flat replace) ``tint`` preserves detail. 0/False = off. Mutually
-//|     exclusive with shadow/flash/dither."""
-static mp_obj_t sprite_get_tint(mp_obj_t self_in) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    return MP_OBJ_NEW_SMALL_INT((s->flags & PICOGAME_SPR_TINT) ? s->flash_color : 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_tint_obj, sprite_get_tint);
-static mp_obj_t sprite_set_tint(mp_obj_t self_in, mp_obj_t v) {
-    picogame_sprite_obj_t *s = MP_OBJ_TO_PTR(self_in);
-    if (!mp_obj_is_true(v)) {
-        set_effect(s, PICOGAME_SPR_TINT, false);
-    } else {
-        s->flash_color = (uint16_t)mp_obj_get_int(v);   // shared colour field with flash
-        set_effect(s, PICOGAME_SPR_TINT, true);
-    }
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_tint_obj, sprite_set_tint);
-MP_PROPERTY_GETSET(sprite_tint_obj, (mp_obj_t)&sprite_get_tint_obj, (mp_obj_t)&sprite_set_tint_obj);
-
-//|     transpose: bool
-//|     """Swap the sprite's X/Y axes - a cheap 90deg turn (no shimmer, unlike ``angle``).
-//|     Combined with ``flip_x``/``flip_y`` it gives all 8 orientations for free. Only on the fast
-//|     path (scale 1.0, angle 0); for rotation WITH scaling use ``angle``. The drawn footprint
-//|     swaps width/height."""
-static mp_obj_t sprite_get_transpose(mp_obj_t self_in) {
-    return mp_obj_new_bool((((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->flags & PICOGAME_SPR_TRANSPOSE) != 0);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_transpose_obj, sprite_get_transpose);
-static mp_obj_t sprite_set_transpose(mp_obj_t self_in, mp_obj_t v) {
-    set_flag(MP_OBJ_TO_PTR(self_in), PICOGAME_SPR_TRANSPOSE, mp_obj_is_true(v));
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_transpose_obj, sprite_set_transpose);
-MP_PROPERTY_GETSET(sprite_transpose_obj, (mp_obj_t)&sprite_get_transpose_obj, (mp_obj_t)&sprite_set_transpose_obj);
-
-//|     data: Any
-//|     """Arbitrary per-sprite user payload for game state (default None)."""
-static mp_obj_t sprite_get_data(mp_obj_t self_in) {
-    return ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->data;
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_data_obj, sprite_get_data);
-static mp_obj_t sprite_set_data(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->data = v;
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_data_obj, sprite_set_data);
-MP_PROPERTY_GETSET(sprite_data_obj, (mp_obj_t)&sprite_get_data_obj, (mp_obj_t)&sprite_set_data_obj);
-
-//|     bitmap: Bitmap
-//|     """The sprite's source bitmap. Assigning a new one swaps graphics and may
-//|     change size; the scene repaints both the old and new bounds next refresh
-//|     (e.g. powerups, resizable HUD bars, text labels)."""
-static mp_obj_t sprite_get_bitmap(mp_obj_t self_in) {
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    return self->bitmap != NULL ? MP_OBJ_FROM_PTR(self->bitmap) : mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_bitmap_obj, sprite_get_bitmap);
-static mp_obj_t sprite_set_bitmap(mp_obj_t self_in, mp_obj_t v) {
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    mp_obj_t bm = mp_arg_validate_type(v, &picogame_bitmap_type, MP_QSTR_bitmap);
-    self->bitmap = MP_OBJ_TO_PTR(bm);
-    self->xf_valid = 0;                              // affine cache depends on bitmap dims
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_bitmap_obj, sprite_set_bitmap);
-MP_PROPERTY_GETSET(sprite_bitmap_obj, (mp_obj_t)&sprite_get_bitmap_obj, (mp_obj_t)&sprite_set_bitmap_obj);
-
-//|     anchor: Tuple[float, float]
-//|     """Pivot as fractions of the bitmap size: ``(0, 0)`` = top-left (default),
-//|     ``(0.5, 0.5)`` = center, ``(0.5, 1.0)`` = bottom-center. ``x``/``y`` then
-//|     refer to this point, so rotating frames or swapping to a different size
-//|     stays aligned. Stored in 1/256 steps."""
-//|
-static mp_obj_t sprite_get_anchor(mp_obj_t self_in) {
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    mp_obj_t t[2] = {
-        mp_obj_new_float(self->anchor_x * (mp_float_t)(1.0 / 256.0)),
-        mp_obj_new_float(self->anchor_y * (mp_float_t)(1.0 / 256.0)),
-    };
-    return mp_obj_new_tuple(2, t);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_get_anchor_obj, sprite_get_anchor);
-static int anchor_to_fp(mp_obj_t o) {
-    mp_float_t f = mp_obj_get_float(o);
-    int v = pg_round_fp8(f);
-    return v < 0 ? 0 : (v > 256 ? 256 : v);
-}
-static mp_obj_t sprite_set_anchor(mp_obj_t self_in, mp_obj_t v) {
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    size_t len;
-    mp_obj_t *items;
-    mp_obj_get_array(v, &len, &items);
-    mp_arg_validate_length(len, 2, MP_QSTR_anchor);
-    self->anchor_x = anchor_to_fp(items[0]);
-    self->anchor_y = anchor_to_fp(items[1]);
-    self->xf_valid = 0;                              // affine cache depends on the pivot
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sprite_set_anchor_obj, sprite_set_anchor);
-MP_PROPERTY_GETSET(sprite_anchor_obj, (mp_obj_t)&sprite_get_anchor_obj, (mp_obj_t)&sprite_set_anchor_obj);
-
-//|     def move(self, x: int, y: int) -> None:
-//|         """Set the sprite position."""
-//|         ...
-//|
-static mp_obj_t sprite_move(mp_obj_t self_in, mp_obj_t x_in, mp_obj_t y_in) {
-    picogame_sprite_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    self->x = obj_to_fp(x_in);
-    self->y = obj_to_fp(y_in);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_3(sprite_move_obj, sprite_move);
-
-//|     def touch(self) -> None:
-//|         """Force this sprite to repaint on the next ``Scene.refresh()`` even though none
-//|         of its tracked properties (position, frame, scale, angle, bitmap) changed. Call
-//|         it after mutating the sprite's bitmap pixels IN PLACE (e.g. streaming a new frame
-//|         into the same buffer), which the dirty-rect tracker can't otherwise detect."""
-//|         ...
-//|
-static mp_obj_t sprite_touch(mp_obj_t self_in) {
-    ((picogame_sprite_obj_t *)MP_OBJ_TO_PTR(self_in))->seq++;
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sprite_touch_obj, sprite_touch);
-
-// Fill (x1,y1,x2,y2) from a Sprite (its drawn aabb), a (x,y) point, or a (x1,y1,x2,y2) rect.
-static void pg_obj_to_box(mp_obj_t o, int *x1, int *y1, int *x2, int *y2) {
-    if (mp_obj_is_type(o, &picogame_sprite_type)) {
-        picogame_sprite_aabb(MP_OBJ_TO_PTR(o), x1, y1, x2, y2);
-        return;
-    }
-    if (mp_obj_is_type(o, &mp_type_tuple) || mp_obj_is_type(o, &mp_type_list)) {
-        size_t len;
-        mp_obj_t *items;
-        mp_obj_get_array(o, &len, &items);
-        if (len == 2) {                       // a point -> a zero-size box
-            *x1 = *x2 = mp_obj_get_int(items[0]);
-            *y1 = *y2 = mp_obj_get_int(items[1]);
-            return;
-        }
-        if (len == 4) {                       // a rect
-            *x1 = mp_obj_get_int(items[0]);
-            *y1 = mp_obj_get_int(items[1]);
-            *x2 = mp_obj_get_int(items[2]);
-            *y2 = mp_obj_get_int(items[3]);
-            return;
-        }
-    }
-    mp_raise_msg_varg(&mp_type_TypeError, MP_ERROR_TEXT("%q must be of type %q or %q, not %q"),
-        MP_QSTR_other, MP_QSTR_Sprite, MP_QSTR_tuple, mp_obj_get_type(o)->name);
-}
-
-//|     def overlaps(self, other: "Sprite | tuple", inset: int = 0) -> bool:
-//|         """True if this sprite's drawn box overlaps ``other`` - an inclusive AABB, so they
-//|         collide the moment they touch. ``other`` may be another Sprite, a point ``(x, y)``,
-//|         or a rect ``(x1, y1, x2, y2)`` (e.g. a trigger zone or the screen for culling).
-//|         The box is anchor/scale/rotation aware. ``inset`` shrinks THIS sprite's box by N px
-//|         on each side, for a fair hitbox smaller than the art."""
-//|         ...
-//|
-static mp_obj_t sprite_overlaps(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
-    enum { ARG_other, ARG_inset };
-    static const mp_arg_t allowed[] = {
-        { MP_QSTR_other, MP_ARG_REQUIRED | MP_ARG_OBJ },
-        { MP_QSTR_inset, MP_ARG_INT, {.u_int = 0} },
-    };
-    mp_arg_val_t args[MP_ARRAY_SIZE(allowed)];
-    mp_arg_parse_all(n_args - 1, pos_args + 1, kw_args, MP_ARRAY_SIZE(allowed), allowed, args);
-    int ax1, ay1, ax2, ay2, bx1, by1, bx2, by2;
-    picogame_sprite_aabb(MP_OBJ_TO_PTR(pos_args[0]), &ax1, &ay1, &ax2, &ay2);
-    pg_obj_to_box(args[ARG_other].u_obj, &bx1, &by1, &bx2, &by2);
-    int in = args[ARG_inset].u_int;                        // inset shrinks the CALLER's box (kw or positional)
-    bool hit = ((ax1 + in) <= bx2) && ((ax2 - in) >= bx1) &&
-        ((ay1 + in) <= by2) && ((ay2 - in) >= by1);
-    return mp_obj_new_bool(hit);
-}
-static MP_DEFINE_CONST_FUN_OBJ_KW(sprite_overlaps_obj, 2, sprite_overlaps);
-
-//|     def near(self, other: "Sprite | tuple", r: int) -> bool:
-//|         """True if this sprite's centre is within ``r`` pixels of ``other``'s centre (squared
-//|         distance, no sqrt) - the round/forgiving test for bullets, pickups, explosions.
-//|         ``other`` may be a Sprite or a point ``(x, y)``. Centres come from the drawn box, so
-//|         it is anchor aware."""
-//|         ...
-//|
-//|
-static mp_obj_t sprite_near(mp_obj_t self_in, mp_obj_t other_in, mp_obj_t r_in) {
-    int ax1, ay1, ax2, ay2;
-    picogame_sprite_aabb(MP_OBJ_TO_PTR(self_in), &ax1, &ay1, &ax2, &ay2);
-    int acx = (ax1 + ax2) / 2, acy = (ay1 + ay2) / 2, bcx, bcy;
-    if (mp_obj_is_type(other_in, &picogame_sprite_type)) {
-        int bx1, by1, bx2, by2;
-        picogame_sprite_aabb(MP_OBJ_TO_PTR(other_in), &bx1, &by1, &bx2, &by2);
-        bcx = (bx1 + bx2) / 2;
-        bcy = (by1 + by2) / 2;
-    } else {
-        size_t len;
-        mp_obj_t *items;
-        mp_obj_get_array(other_in, &len, &items);
-        if (len != 2) {
-            mp_raise_msg_varg(&mp_type_TypeError, MP_ERROR_TEXT("%q must be of type %q or %q, not %q"),
-                MP_QSTR_other, MP_QSTR_Sprite, MP_QSTR_tuple, mp_obj_get_type(other_in)->name);
-        }
-        bcx = mp_obj_get_int(items[0]);
-        bcy = mp_obj_get_int(items[1]);
-    }
-    mp_int_t r = mp_obj_get_int(r_in), dx = acx - bcx, dy = acy - bcy;
-    return mp_obj_new_bool(dx * dx + dy * dy < r * r);
-}
-static MP_DEFINE_CONST_FUN_OBJ_3(sprite_near_obj, sprite_near);
-
-static const mp_rom_map_elem_t picogame_sprite_locals_dict_table[] = {
-    { MP_ROM_QSTR(MP_QSTR_x), MP_ROM_PTR(&sprite_x_obj) },
-    { MP_ROM_QSTR(MP_QSTR_y), MP_ROM_PTR(&sprite_y_obj) },
-    { MP_ROM_QSTR(MP_QSTR_fx), MP_ROM_PTR(&sprite_fx_obj) },
-    { MP_ROM_QSTR(MP_QSTR_fy), MP_ROM_PTR(&sprite_fy_obj) },
-    { MP_ROM_QSTR(MP_QSTR_frame), MP_ROM_PTR(&sprite_frame_obj) },
-    { MP_ROM_QSTR(MP_QSTR_visible), MP_ROM_PTR(&sprite_visible_obj) },
-    { MP_ROM_QSTR(MP_QSTR_flip_x), MP_ROM_PTR(&sprite_flip_x_obj) },
-    { MP_ROM_QSTR(MP_QSTR_flip_y), MP_ROM_PTR(&sprite_flip_y_obj) },
-    { MP_ROM_QSTR(MP_QSTR_scale), MP_ROM_PTR(&sprite_scale_obj) },
-    { MP_ROM_QSTR(MP_QSTR_angle), MP_ROM_PTR(&sprite_angle_obj) },
-    { MP_ROM_QSTR(MP_QSTR_shadow), MP_ROM_PTR(&sprite_shadow_obj) },
-    { MP_ROM_QSTR(MP_QSTR_flash), MP_ROM_PTR(&sprite_flash_obj) },
-    { MP_ROM_QSTR(MP_QSTR_dither), MP_ROM_PTR(&sprite_dither_obj) },
-    { MP_ROM_QSTR(MP_QSTR_tint), MP_ROM_PTR(&sprite_tint_obj) },
-    { MP_ROM_QSTR(MP_QSTR_transpose), MP_ROM_PTR(&sprite_transpose_obj) },
-    { MP_ROM_QSTR(MP_QSTR_data), MP_ROM_PTR(&sprite_data_obj) },
-    { MP_ROM_QSTR(MP_QSTR_bitmap), MP_ROM_PTR(&sprite_bitmap_obj) },
-    { MP_ROM_QSTR(MP_QSTR_anchor), MP_ROM_PTR(&sprite_anchor_obj) },
-    { MP_ROM_QSTR(MP_QSTR_overlaps), MP_ROM_PTR(&sprite_overlaps_obj) },
-    { MP_ROM_QSTR(MP_QSTR_near), MP_ROM_PTR(&sprite_near_obj) },
-    { MP_ROM_QSTR(MP_QSTR_move), MP_ROM_PTR(&sprite_move_obj) },
-    { MP_ROM_QSTR(MP_QSTR_touch), MP_ROM_PTR(&sprite_touch_obj) },
-};
-static MP_DEFINE_CONST_DICT(picogame_sprite_locals_dict, picogame_sprite_locals_dict_table);
-
-MP_DEFINE_CONST_OBJ_TYPE(
-    picogame_sprite_type,
-    MP_QSTR_Sprite,
-    MP_TYPE_FLAG_HAS_SPECIAL_ACCESSORS,
-    make_new, picogame_sprite_make_new,
-    locals_dict, &picogame_sprite_locals_dict
-    );
-
-// ---------------------------------------------------------------------------
-// StripDraw  (immediate-mode draw layer; struct in shared-module/picogame/__init__.h)
-// ---------------------------------------------------------------------------
-
-//| class StripDraw:
-//|     """An immediate-mode draw layer that holds NO pixel buffer. Each refresh, for
-//|     every render strip overlapping its rect, ``callback(view, vx, vy, vw, vh)`` is
-//|     called with a :py:class:`Canvas` ``view`` pointing straight at the live strip
-//|     buffer - so you draw primitives directly into the frame (zero RAM, vs a Canvas
-//|     which costs width*height*2 bytes). The view's local (0, 0) is screen pixel
-//|     (vx, vy); (vw, vh) is the strip size. Draw only the rows in [vy, vy+vh) for
-//|     speed (anything outside the view is clipped anyway). The rect is repainted every
-//|     frame, so use it for animated / scanline content (pseudo-3D, gradients,
-//|     procedural backgrounds), not static art (use Canvas for that).
-//|
-//|     COORDINATE CONTRACT: ``vx`` is the RENDER REGION's origin (NOT this layer's x), and the
-//|     view spans the FULL region WIDTH (the layer's rect only gates which ROWS run). So draw at
-//|     ABSOLUTE screen coords minus (vx, vy), and fill only your own rect with ``fill_rect`` -
-//|     ``view.clear()`` fills the whole region width. (When you render a StripDraw via
-//|     ``picogame.render([sd], buf, x,y,x+w,y+h)`` the region == the rect, so vx == x.)
-//|     Text via ``Canvas.text`` is ASCII (the built-in font); non-ASCII has no glyph."""
-//|
-//|     def __init__(
-//|         self,
-//|         callback: Callable[[Canvas, int, int, int, int], None],
-//|         x: int = 0,
-//|         y: int = 0,
-//|         width: int = 0,
-//|         height: int = 0,
-//|     ) -> None: ...
-//|
-static mp_obj_t picogame_stripdraw_make_new(const mp_obj_type_t *type, size_t n_args,
-    size_t n_kw, const mp_obj_t *all_args) {
-    enum { ARG_callback, ARG_x, ARG_y, ARG_width, ARG_height, ARG_always_dirty };
-    static const mp_arg_t allowed_args[] = {
-        { MP_QSTR_callback, MP_ARG_REQUIRED | MP_ARG_OBJ },
-        { MP_QSTR_x, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_y, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_width, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_height, MP_ARG_INT, {.u_int = 0} },
-        { MP_QSTR_always_dirty, MP_ARG_BOOL | MP_ARG_KW_ONLY, {.u_bool = true} },
-    };
-    mp_arg_val_t args[MP_ARRAY_SIZE(allowed_args)];
-    mp_arg_parse_all_kw_array(n_args, n_kw, all_args, MP_ARRAY_SIZE(allowed_args), allowed_args, args);
-
-    picogame_stripdraw_obj_t *self = mp_obj_malloc(picogame_stripdraw_obj_t, type);
-    self->callback = args[ARG_callback].u_obj;
-    self->x = args[ARG_x].u_int;
-    self->y = args[ARG_y].u_int;
-    self->w = args[ARG_width].u_int;
-    self->h = args[ARG_height].u_int;
-    self->faulted = false;
-    self->always_dirty = args[ARG_always_dirty].u_bool;
-    picogame_dirty_reset(&self->dx1);        // render once on first refresh (even when always_dirty=False)
-    picogame_dirty_union(&self->dx1, self->x, self->y, self->x + self->w, self->y + self->h);
-    // A buffer-less Canvas reused as the per-strip drawing view: its `data` is
-    // repointed at the live strip each blit, so no surface RAM is allocated here.
-    picogame_canvas_obj_t *view = mp_obj_malloc(picogame_canvas_obj_t, &picogame_canvas_type);
-    view->data = NULL;
-    view->data_obj = MP_OBJ_NULL;
-    view->w = 0;
-    view->h = 0;
-    view->x = 0;
-    view->y = 0;
-    view->transparent = 0;
-    view->has_transparent = false;
-    picogame_canvas_dirty_reset(view);
-    self->view = MP_OBJ_FROM_PTR(view);
-    return MP_OBJ_FROM_PTR(self);
-}
-
-//|
-//|     x: int
-//|     y: int
-//|     width: int
-//|     height: int
-//|     """The screen rect repainted each refresh (read/write). Move or resize the layer
-//|     by assigning these. Shrinking the rect leaves stale pixels behind - follow a
-//|     shrink with ``scene.invalidate()`` for a clean repaint (as the fx helpers do)."""
-static mp_obj_t sd_get_x(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->x);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sd_get_x_obj, sd_get_x);
-static mp_obj_t sd_set_x(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->x = mp_obj_get_int(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sd_set_x_obj, sd_set_x);
-MP_PROPERTY_GETSET(sd_x_obj, (mp_obj_t)&sd_get_x_obj, (mp_obj_t)&sd_set_x_obj);
-
-static mp_obj_t sd_get_y(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->y);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sd_get_y_obj, sd_get_y);
-static mp_obj_t sd_set_y(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->y = mp_obj_get_int(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sd_set_y_obj, sd_set_y);
-MP_PROPERTY_GETSET(sd_y_obj, (mp_obj_t)&sd_get_y_obj, (mp_obj_t)&sd_set_y_obj);
-
-static mp_obj_t sd_get_width(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->w);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sd_get_width_obj, sd_get_width);
-static mp_obj_t sd_set_width(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->w = mp_obj_get_int(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sd_set_width_obj, sd_set_width);
-MP_PROPERTY_GETSET(sd_width_obj, (mp_obj_t)&sd_get_width_obj, (mp_obj_t)&sd_set_width_obj);
-
-static mp_obj_t sd_get_height(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->h);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sd_get_height_obj, sd_get_height);
-static mp_obj_t sd_set_height(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->h = mp_obj_get_int(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sd_set_height_obj, sd_set_height);
-MP_PROPERTY_GETSET(sd_height_obj, (mp_obj_t)&sd_get_height_obj, (mp_obj_t)&sd_set_height_obj);
-
-//|     always_dirty: bool
-//|     """True (default): repaint every frame - for animated content (pseudo-3D, gradients). False:
-//|     repaint only after an ``invalidate()`` call (or when overlapped by another dirty layer) - for on-change UI,
-//|     so a static panel doesn't re-rasterize+re-push every frame. With False you MUST invalidate() on
-//|     every content/visibility change (it's invisible until you do)."""
-//|
-static mp_obj_t sd_get_always_dirty(mp_obj_t self_in) {
-    return mp_obj_new_bool(((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->always_dirty);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(sd_get_always_dirty_obj, sd_get_always_dirty);
-static mp_obj_t sd_set_always_dirty(mp_obj_t self_in, mp_obj_t v) {
-    ((picogame_stripdraw_obj_t *)MP_OBJ_TO_PTR(self_in))->always_dirty = mp_obj_is_true(v);
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(sd_set_always_dirty_obj, sd_set_always_dirty);
-MP_PROPERTY_GETSET(sd_always_dirty_obj, (mp_obj_t)&sd_get_always_dirty_obj, (mp_obj_t)&sd_set_always_dirty_obj);
-
-//|     def invalidate(self, x: int = 0, y: int = 0, w: int = 0, h: int = 0) -> None:
-//|         """Mark dirty so the layer repaints on the next refresh (only needed when
-//|         ``always_dirty=False``). With no args, the whole layer repaints. Pass a rect in
-//|         VIEW-LOCAL coordinates (the same (0,0)-at-``(vx, vy)`` space the draw callback uses) to
-//|         repaint only that region - like Canvas/Tilemap, the Scene then recomposites and pushes just
-//|         those rows. Repeated calls union; the rect is clamped to the layer."""
-//|
-//|
-static mp_obj_t sd_invalidate(size_t n_args, const mp_obj_t *args) {
-    picogame_stripdraw_obj_t *self = MP_OBJ_TO_PTR(args[0]);
-    if (n_args >= 5) {                       // (x, y, w, h) in view-local coords -> clamped scene rect
-        int x1 = self->x + mp_obj_get_int(args[1]);
-        int y1 = self->y + mp_obj_get_int(args[2]);
-        int x2 = x1 + mp_obj_get_int(args[3]);
-        int y2 = y1 + mp_obj_get_int(args[4]);
-        if (x1 < self->x) {
-            x1 = self->x;
-        }
-        if (y1 < self->y) {
-            y1 = self->y;
-        }
-        if (x2 > self->x + self->w) {
-            x2 = self->x + self->w;
-        }
-        if (y2 > self->y + self->h) {
-            y2 = self->y + self->h;
-        }
-        if (x2 > x1 && y2 > y1) {
-            picogame_dirty_union(&self->dx1, x1, y1, x2, y2);
-        }
-    } else {                                 // whole layer
-        picogame_dirty_union(&self->dx1, self->x, self->y, self->x + self->w, self->y + self->h);
-    }
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_VAR_BETWEEN(sd_invalidate_obj, 1, 5, sd_invalidate);
-
-static const mp_rom_map_elem_t picogame_stripdraw_locals_dict_table[] = {
-    { MP_ROM_QSTR(MP_QSTR_x), MP_ROM_PTR(&sd_x_obj) },
-    { MP_ROM_QSTR(MP_QSTR_y), MP_ROM_PTR(&sd_y_obj) },
-    { MP_ROM_QSTR(MP_QSTR_width), MP_ROM_PTR(&sd_width_obj) },
-    { MP_ROM_QSTR(MP_QSTR_height), MP_ROM_PTR(&sd_height_obj) },
-    { MP_ROM_QSTR(MP_QSTR_always_dirty), MP_ROM_PTR(&sd_always_dirty_obj) },
-    { MP_ROM_QSTR(MP_QSTR_invalidate), MP_ROM_PTR(&sd_invalidate_obj) },
-};
-static MP_DEFINE_CONST_DICT(picogame_stripdraw_locals_dict, picogame_stripdraw_locals_dict_table);
-
-MP_DEFINE_CONST_OBJ_TYPE(
-    picogame_stripdraw_type,
-    MP_QSTR_StripDraw,
-    MP_TYPE_FLAG_HAS_SPECIAL_ACCESSORS,
-    make_new, picogame_stripdraw_make_new,
-    locals_dict, &picogame_stripdraw_locals_dict
-    );
-
-//| class Triangles:
-//|     def __init__(self, verts: ReadableBuffer, colors: ReadableBuffer) -> None:
-//|         """A retained SCREEN-SPACE triangle batch drawn entirely in C by the compositor:
-//|         ``verts`` = int16 x0,y0,x1,y1,x2,y2 per triangle, ``colors`` = uint16 wire RGB565 per
-//|         triangle - both CALLER-OWNED (fill them in place each frame). Set ``count`` to how
-//|         many triangles should draw; the assignment marks the layer dirty (full screen).
-//|         Unlike a StripDraw callback this runs no Python per strip, and unlike a Canvas it
-//|         holds no pixel buffer - the batch rasterises straight into each render strip with
-//|         a cheap band reject. The 3D-scene layer: pg.project into the arrays, painter's-order
-//|         the faces, set count, scene.refresh()."""
-//|         ...
-//|
-static mp_obj_t picogame_triangles_make_new(const mp_obj_type_t *type, size_t n_args,
-    size_t n_kw, const mp_obj_t *all_args) {
-    mp_arg_check_num(n_args, n_kw, 2, 2, false);
-    picogame_triangles_obj_t *self = mp_obj_malloc(picogame_triangles_obj_t, type);
-    mp_buffer_info_t vi, ci;
-    mp_get_buffer_raise(all_args[0], &vi, MP_BUFFER_READ);
-    mp_get_buffer_raise(all_args[1], &ci, MP_BUFFER_READ);
-    self->verts_obj = all_args[0];
-    self->colors_obj = all_args[1];
-    self->verts = (const int16_t *)vi.buf;
-    self->colors = (const uint16_t *)ci.buf;
-    size_t cap_v = vi.len / 12;                    // 6 int16 = 12 bytes per triangle
-    size_t cap_c = ci.len >> 1;
-    self->cap = (uint16_t)(cap_v < cap_c ? cap_v : cap_c);
-    self->count = 0;
-    picogame_dirty_reset(&self->dx1);
-    return MP_OBJ_FROM_PTR(self);
-}
-
-//|     count: int
-//|     """How many triangles of the batch draw next refresh (clamped to the buffer
-//|     capacity). Assigning marks the layer dirty for a full repaint."""
-//|
-//|
-static mp_obj_t tri_get_count(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_triangles_obj_t *)MP_OBJ_TO_PTR(self_in))->count);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(tri_get_count_obj, tri_get_count);
-static mp_obj_t tri_set_count(mp_obj_t self_in, mp_obj_t v) {
-    picogame_triangles_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    int n = mp_obj_get_int(v);
-    if (n < 0) {
-        n = 0;
-    }
-    if (n > self->cap) {
-        n = self->cap;
-    }
-    self->count = (uint16_t)n;
-    picogame_dirty_union(&self->dx1, 0, 0, 32767, 32767);   // clipped to the play rect later
-    return mp_const_none;
-}
-static MP_DEFINE_CONST_FUN_OBJ_2(tri_set_count_obj, tri_set_count);
-MP_PROPERTY_GETSET(tri_count_obj, (mp_obj_t)&tri_get_count_obj, (mp_obj_t)&tri_set_count_obj);
-
-static const mp_rom_map_elem_t picogame_triangles_locals_dict_table[] = {
-    { MP_ROM_QSTR(MP_QSTR_count), MP_ROM_PTR(&tri_count_obj) },
-};
-static MP_DEFINE_CONST_DICT(picogame_triangles_locals_dict, picogame_triangles_locals_dict_table);
-
-MP_DEFINE_CONST_OBJ_TYPE(
-    picogame_triangles_type,
-    MP_QSTR_Triangles,
-    MP_TYPE_FLAG_HAS_SPECIAL_ACCESSORS,
-    make_new, picogame_triangles_make_new,
-    locals_dict, &picogame_triangles_locals_dict
-    );
-
-// ---------------------------------------------------------------------------
-// Module-level functions
-// ---------------------------------------------------------------------------
 
 //| """2D game engine for the PicoPad and similar boards.
 //|
 //| Draws arbitrary-size sprites (unlike ``_stage``'s fixed 16x16 tiles) to a
 //| ``busdisplay`` through a reusable strip buffer, with a dirty-rect scene,
-//| tilemaps, particles, a drawing canvas and camera/effects."""
+//| tilemaps, particles, a drawing canvas and camera/effects.
+//|
+//| **Relationship to displayio.** picogame does not replace or extend
+//| :py:mod:`displayio`; it is a second, game-shaped way to drive the same
+//| hardware, and the two do not share a display at the same time. displayio
+//| retains a widget tree the supervisor refreshes for you, holds a full
+//| :py:class:`~displayio.Bitmap` per image and is the right tool for UI. picogame
+//| composites on demand, one horizontal strip at a time, into a buffer the game
+//| owns: nothing is retained per pixel, so a scrolling game fits in the RAM a
+//| microcontroller actually has, and the game decides when a frame happens
+//| (``scene.refresh()``).
+//|
+//| The seam is the display object. A ``picogame.Scene`` takes the same
+//| :py:class:`~busdisplay.BusDisplay` displayio uses - it just talks to it
+//| directly instead of through the displayio refresh loop, so set
+//| ``display.auto_refresh = False`` (``picogame_game.setup()`` does this) and let
+//| the game drive. On boards that scan out of RAM, ``picogame.Framebuffer`` takes
+//| that buffer instead. Bitmaps are separate types: displayio's is a mutable
+//| indexed surface, picogame's is read-only pixel data (PAL8 or wire RGB565) that
+//| may live in flash, so it costs no RAM at all."""
 //|
 //| RGB565: int
 //| """16-bit color bitmap format (wire byte order)."""
@@ -1407,7 +409,7 @@ uint8_t picogame_kind_of(mp_obj_t o) {
     if (mp_obj_is_type(o, &picogame_triangles_type)) {
         return PICOGAME_KIND_TRIANGLES;
     }
-    mp_raise_TypeError(MP_ERROR_TEXT("expected a Sprite, Tilemap, Particles, Canvas, StripDraw or Triangles"));
+    mp_arg_error_invalid(MP_QSTR_item);        // the accepted layer types are listed in the docs
 }
 
 static mp_obj_t picogame_render_fun(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
@@ -1609,7 +611,7 @@ static const mp_arg_t pg_fbm2d_args[] = {
 };
 
 #if 0   // float reference fbm - superseded by the fixed-point path below
-//| def fbm2d(x, y, *, octaves=4, seed=0, lacunarity=2.0, gain=0.5) -> float: ...
+//| def fbm2d(x: float, y: float, *, octaves: int = 4, seed: int = 0, lacunarity: float = 2.0, gain: float = 0.5) -> float: ...
 static mp_obj_t picogame_fbm2d(size_t n_args, const mp_obj_t *pos, mp_map_t *kw) {
     mp_arg_val_t a[6];
     mp_arg_parse_all(n_args, pos, kw, 6, pg_fbm2d_args, a);
@@ -1629,7 +631,7 @@ static mp_obj_t picogame_fbm2d(size_t n_args, const mp_obj_t *pos, mp_map_t *kw)
 }
 static MP_DEFINE_CONST_FUN_OBJ_KW(picogame_fbm2d_obj, 2, picogame_fbm2d);
 
-//| def fbm1d(x, *, octaves=4, seed=0, lacunarity=2.0, gain=0.5) -> float: ...
+//| def fbm1d(x: float, *, octaves: int = 4, seed: int = 0, lacunarity: float = 2.0, gain: float = 0.5) -> float: ...
 //|
 //|
 static mp_obj_t picogame_fbm1d(size_t n_args, const mp_obj_t *pos, mp_map_t *kw) {
@@ -1741,133 +743,6 @@ static mp_obj_t picogame_fbm1d_fx(size_t n_args, const mp_obj_t *pos, mp_map_t *
 }
 static MP_DEFINE_CONST_FUN_OBJ_KW(picogame_fbm1d_fx_obj, 1, picogame_fbm1d_fx);
 
-#if CIRCUITPY_PICOGAME_FRAMEBUFFER
-// ---------------------------------------------------------------------------
-// Framebuffer  (a RAM render target used in place of a BusDisplay; scanout-buffer
-// platforms - WASM playground, desktop sim, FruitJam DVI/HSTX)
-// ---------------------------------------------------------------------------
-//| class Framebuffer:
-//|     """A RAM framebuffer render target that a Scene or :py:func:`render` can draw
-//|     into instead of a BusDisplay. ``buffer`` must be a writable buffer of at least
-//|     ``width*height*2`` bytes (``width*height`` for ``rgb332=True``); the caller owns it
-//|     (a ``bytearray`` in the browser, the DVI scanout buffer on FruitJam). By default the
-//|     pixels are wire-order RGB565 (the engine's internal format); ``native_rgb565=True``
-//|     byte-swaps each finished region to NATIVE RGB565 - the format 16-bit picodvi /
-//|     canvas scanout targets expect; ``rgb332=True`` quantizes each finished region to
-//|     RGB332 bytes - the format of 8-bit picodvi scanout (FruitJam 640x480, which the
-//|     hardware only offers at 8bpp). Assets, palettes and ``rgb565()`` stay wire-order
-//|     RGB565 throughout regardless of the output format."""
-//|
-//|     def __init__(
-//|         self,
-//|         buffer: WriteableBuffer,
-//|         width: int,
-//|         height: int,
-//|         *,
-//|         native_rgb565: bool = False,
-//|         rgb332: bool = False,
-//|     ) -> None: ...
-//|
-//|
-// Static SRAM compose strip (see the scratch comment in make_new). 640*16*2 = 20 KB .bss,
-// only on CIRCUITPY_PICOGAME_FRAMEBUFFER builds (fb boards have the SRAM to spare).
-#define PICOGAME_FB_SCRATCH_MAX_W 640
-static uint16_t picogame_fb_scratch_sram[PICOGAME_FB_SCRATCH_MAX_W * PICOGAME_FB_SCRATCH_H];
-
-static mp_obj_t picogame_framebuffer_make_new(const mp_obj_type_t *type, size_t n_args,
-    size_t n_kw, const mp_obj_t *all_args) {
-    enum { ARG_buffer, ARG_width, ARG_height, ARG_native_rgb565, ARG_rgb332 };
-    static const mp_arg_t allowed_args[] = {
-        { MP_QSTR_buffer, MP_ARG_REQUIRED | MP_ARG_OBJ },
-        { MP_QSTR_width, MP_ARG_REQUIRED | MP_ARG_INT },
-        { MP_QSTR_height, MP_ARG_REQUIRED | MP_ARG_INT },
-        { MP_QSTR_native_rgb565, MP_ARG_KW_ONLY | MP_ARG_BOOL, {.u_bool = false} },
-        { MP_QSTR_rgb332, MP_ARG_KW_ONLY | MP_ARG_BOOL, {.u_bool = false} },
-    };
-    mp_arg_val_t args[MP_ARRAY_SIZE(allowed_args)];
-    mp_arg_parse_all_kw_array(n_args, n_kw, all_args, MP_ARRAY_SIZE(allowed_args), allowed_args, args);
-    mp_int_t width = mp_arg_validate_int_range(args[ARG_width].u_int, 1, 4096, MP_QSTR_width);
-    mp_int_t height = mp_arg_validate_int_range(args[ARG_height].u_int, 1, 4096, MP_QSTR_height);
-    if (args[ARG_native_rgb565].u_bool && args[ARG_rgb332].u_bool) {
-        mp_arg_error_invalid(MP_QSTR_format);          // native_rgb565 and rgb332 are exclusive
-    }
-    bool rgb332 = args[ARG_rgb332].u_bool;
-
-    mp_buffer_info_t bi;
-    mp_get_buffer_raise(args[ARG_buffer].u_obj, &bi, MP_BUFFER_WRITE);
-    uint64_t need = (uint64_t)width * (uint64_t)height * (rgb332 ? 1u : 2u);
-    if ((uint64_t)bi.len < need) {
-        mp_raise_ValueError(MP_ERROR_TEXT("buffer too small"));
-    }
-
-    picogame_framebuffer_obj_t *self = mp_obj_malloc(picogame_framebuffer_obj_t, type);
-    self->buffer = args[ARG_buffer].u_obj;
-    self->fb = (uint16_t *)bi.buf;
-    self->width = width;
-    self->height = height;
-    self->fmt = rgb332 ? PICOGAME_FB_RGB332
-        : (args[ARG_native_rgb565].u_bool ? PICOGAME_FB_NATIVE565 : PICOGAME_FB_WIRE565);
-    // A LIVE scanout buffer (picodvi/HDMI) is read continuously, so picogame_render_framebuffer
-    // composes each band into this PRIVATE strip and only memcpys the FINISHED band into the fb.
-    // That serves BOTH targets: (a) native -> also byte-swap the strip so the fb never holds wire
-    // (no pink); (b) wire -> no swap, but the off-screen compose still stops the beam from sampling
-    // a half-composited region (background filled, sprite not yet drawn) = no sprite/HUD flicker.
-    // Always allocated for the FB target; the WASM/sim path (read out after present, not live) just
-    // pays a small strip + one memcpy. See PICOGAME_FB_SCRATCH_H.
-    //
-    // The scratch must be FAST memory: on a PSRAM-heap board (Fruit Jam) a heap bytearray
-    // lands in external PSRAM and every compose write pays QSPI latency (measured 8.7 vs
-    // 64+ MB/s SRAM; a full-res StripDraw frame ballooned refresh to ~30-38 ms). One static
-    // SRAM strip serves every Framebuffer (compose is synchronous) up to 640 px wide; wider
-    // targets fall back to the heap.
-    self->scratch_buf = mp_const_none;
-    self->scratch = NULL;
-    self->scratch_rows = 0;
-    {
-        int rows = PICOGAME_FB_SCRATCH_H;
-        if (rows > height) {
-            rows = height;
-        }
-        if (width <= PICOGAME_FB_SCRATCH_MAX_W) {
-            self->scratch = picogame_fb_scratch_sram;
-        } else {
-            mp_obj_t sb = mp_obj_new_bytearray_of_zeros((size_t)width * (size_t)rows * 2u);
-            mp_buffer_info_t sbi;
-            mp_get_buffer_raise(sb, &sbi, MP_BUFFER_WRITE);
-            self->scratch_buf = sb;
-            self->scratch = (uint16_t *)sbi.buf;
-        }
-        self->scratch_rows = rows;
-    }
-    return MP_OBJ_FROM_PTR(self);
-}
-
-static mp_obj_t picogame_framebuffer_get_width(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_framebuffer_obj_t *)MP_OBJ_TO_PTR(self_in))->width);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(picogame_framebuffer_get_width_obj, picogame_framebuffer_get_width);
-MP_PROPERTY_GETTER(picogame_framebuffer_width_obj, (mp_obj_t)&picogame_framebuffer_get_width_obj);
-
-static mp_obj_t picogame_framebuffer_get_height(mp_obj_t self_in) {
-    return MP_OBJ_NEW_SMALL_INT(((picogame_framebuffer_obj_t *)MP_OBJ_TO_PTR(self_in))->height);
-}
-static MP_DEFINE_CONST_FUN_OBJ_1(picogame_framebuffer_get_height_obj, picogame_framebuffer_get_height);
-MP_PROPERTY_GETTER(picogame_framebuffer_height_obj, (mp_obj_t)&picogame_framebuffer_get_height_obj);
-
-static const mp_rom_map_elem_t picogame_framebuffer_locals_dict_table[] = {
-    { MP_ROM_QSTR(MP_QSTR_width), MP_ROM_PTR(&picogame_framebuffer_width_obj) },
-    { MP_ROM_QSTR(MP_QSTR_height), MP_ROM_PTR(&picogame_framebuffer_height_obj) },
-};
-static MP_DEFINE_CONST_DICT(picogame_framebuffer_locals_dict, picogame_framebuffer_locals_dict_table);
-
-MP_DEFINE_CONST_OBJ_TYPE(
-    picogame_framebuffer_type,
-    MP_QSTR_Framebuffer,
-    MP_TYPE_FLAG_HAS_SPECIAL_ACCESSORS,
-    make_new, picogame_framebuffer_make_new,
-    locals_dict, &picogame_framebuffer_locals_dict
-    );
-#endif // CIRCUITPY_PICOGAME_FRAMEBUFFER
 
 static const mp_rom_map_elem_t picogame_module_globals_table[] = {
     { MP_ROM_QSTR(MP_QSTR___name__), MP_ROM_QSTR(MP_QSTR_picogame) },
